backend-bot/requirements.txt
fastapi==0.81.0
httpx
python-dotenv